+++ src/main/java/kitchenpos/table/dto/TableGroupOrderTableRequest.java
@@ -1,0 +1,13 @@
+package kitchenpos.table.dto;
+
+public class TableGroupOrderTableRequest {
+    private Long id;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+}
